Successfully received a message

+++ include/shared_memory.h
@@ -9,7 +9,7 @@ struct clock {
 };
 
 struct msgbuf {
-    int mtype;
+    long mtype;
     char mtext[MSGSZ];
 };
